Refactor to organize files into different packages
Classes are currently scattered within the sloth package, mixing commands, exceptions, and task domain types.

This makes navigation noisy, blurs package responsibilities, and complicates maintenance and Checkstyle/package rules.

Move command-related classes into sloth.command, exception types into sloth.exception, and task classes into sloth.task. Update package declarations, imports, and references.

Using feature-focused subpackages improves cohesion and discoverability without changing behavior.

No functional modifications; code compiles the same. Directory structure and imports updated only.
