+++ src/main/java/sloth/TaskList.java
@@ -1,5 +1,8 @@
 package sloth;
+
 import java.util.ArrayList;
+import sloth.task.Task;
+import sloth.exception.*;
 
 /**
  * Manages a collection of tasks with operations for adding, removing, and modifying tasks.
