+++ src/main/java/sloth/TaskParser.java
@@ -1,5 +1,21 @@
 package sloth;
 
+import sloth.command.AddTaskCommand;
+import sloth.command.ByeCommand;
+import sloth.command.Command;
+import sloth.command.DeleteCommand;
+import sloth.command.FindCommand;
+import sloth.command.ListCommand;
+import sloth.command.MarkCommand;
+import sloth.command.UnmarkCommand;
+import sloth.task.Task;
+import sloth.task.Deadline;
+import sloth.task.ToDo;
+import sloth.task.Event;
+import sloth.exception.SlothException;
+import sloth.exception.EmptyDescriptionException;
+import sloth.exception.UnknownCommandException;
+import sloth.exception.ParseException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
