+++ src/test/java/sloth/TaskParserTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.DisplayName;
 import static org.junit.jupiter.api.Assertions.*;
 import java.time.LocalDateTime;
+import sloth.exception.SlothException;
 
 public class TaskParserTest {
